TUP-26696 Enhance context propagation over reference project - Propagation https://jira.talendforge.org/browse/TUP-26696
